+++ NetworkTableRelayApp/build.gradle
@@ -5,7 +5,7 @@ plugins {
     id 'edu.wpi.first.WpilibTools' version '1.1.0'
 }
 
-mainClassName = 'Program'
+mainClassName = 'com/networktablerelayapp/Main'
 
 wpilibTools.deps.wpilibVersion = wpi.versions.wpilibVersion.get()
